+++ src/main/java/com/ttubeog/domain/store/dto/response/RegisterStoreRes.java
@@ -28,10 +28,10 @@ public class RegisterStoreRes {
     private String detailAddress;
 
     @Schema(description = "위도")
-    private Float latitude;
+    private Double latitude;
 
     @Schema(description = "경도")
-    private Float longitude;
+    private Double longitude;
 
     @Schema(description = "이미지")
     private String image;
@@ -44,7 +44,7 @@ public class RegisterStoreRes {
 
     @Builder
     public RegisterStoreRes(Long storeId, Long memberId, String name, String info, Long dongAreaId,
-                            String detailAddress,Float latitude, Float longitude, String image, Float stars, StoreType type) {
+                            String detailAddress, Double latitude, Double longitude, String image, Float stars, StoreType type) {
         this.storeId = storeId;
         this.memberId = memberId;
         this.name = name;
